Add a tool to intersect two lists
Add a tool to intersect two lists and output unique lines from the
first, common lines in the intersection, and unique lines from the
other.
Also, add a tool that given a list of lines to exclude will drop them
if present.
These tools are useful for combining new leads into a mailing list and
removing unsubscribes, or various other things.